# [inductor] scale down RBLOCK for occupancy