@shunting314 2023-09-14
Here is a perf test [link](https://hud.pytorch.org/benchmark/compilers?startTime=Thu%2C%2007%20Sep%202023%2018%3A45%3A31%20GMT&stopTime=Thu%2C%2014%20Sep%202023%2018%3A45%3A31%20GMT&granularity=hour&suite=torchbench&mode=training&dtype=amp&lBranch=shunting-scale-down-rblock&lCommit=f3a0be938587a318cb4e4651f8804bf50363da5a&rBranch=main&rCommit=e716505345f36c720232802e7e7a97321917d29a)
- HF 1.80x -> 1.82x .
- TB 1.67x -> 1.68x. It's not like noise since I see 
   - hf_GPT2_large 1.75x -> 1.88x
   - BERT_pytorch: 3.04x -> 3.11x
   - drq: 1.48x -> 1.53x
- TIMM 1.69x -> 1.70x . This can be noise.

Compilation time increases a bit
- HF 83s -> 87s
- TB 67s -> 68s
- TIMM 109s -> 111s

I hope we can trade the compilation time with some perf wins.