@shunting314 2023-09-14
> Why would there be compilation time increases? Can we get runtime improvements without the compilation time increase?

That's because we generate extra configs to compile and benchmark. We could skip benchmarking an old config if we reduce its RBLOCK, but
- that may not change the overall compilation time much since I think compared to compilation, benchmarking is not that slow
- that may cause bad perf for some models. The changed config is not guaranteed to be better than the old config in every case. 

I discussed with Elias offline the idea to make the parallel compilation tasks more finer-grain.  Currently each tasks compiles all configs for a kernel. But maybe let each task compiles a single config can speedup things. Hopefully I can trade some compilation time here