@shunting314 2023-09-14
Hmm, I realized the extra configs are all get compiled by the main process (not any of the child process for parallel compiling). The reason is the parallel compiling subprocess don't have CUDA initialized. We can not load the cubin (in triton) so the number of registers are not available.

Not sure if there is anyway to workaround.