@shunting314 2023-09-14
> The reason is the parallel compiling subprocess don't have CUDA initialized. We can not load the cubin (in triton) so the number of registers are not available.

In theory there is one way to work around:

```
$ ptxas -arch sm_80 sumrow.ptx -v
ptxas info    : 0 bytes gmem
ptxas info    : Compiling entry function 'sumrow_kernel_0d1d23d' for 'sm_80'
ptxas info    : Function properties for sumrow_kernel_0d1d23d
    0 bytes stack frame, 0 bytes spill stores, 0 bytes spill loads
ptxas info    : Used 32 registers, 376 bytes cmem[0]
```

with '-v' ptxas will report the number of registers used.  But in inductor stack, ptxas is called by triton. We'll need to change triton to make num-used-registers information available without loading the cubin.